requirements.txt
attrs==25.3.0
cattrs==24.1.3
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.1.8
decorator==5.2.1
exceptiongroup==1.3.0
flatbuffers==25.2.10
future==1.0.0
geocoder==1.38.1
h11==0.16.0
idna==3.10
jh2==5.0.9
kafka-python==2.2.15
niquests==3.14.1
numpy==1.24.4
openmeteo-requests==1.5.0
openmeteo-sdk==1.20.1
pandas==2.0.3
platformdirs==4.3.6
python-dateutil==2.9.0.post0
pytz==2025.2
qh3==1.5.3
ratelim==0.1.6
requests==2.32.4
requests-cache==1.2.1
retry-requests==2.0.0
six==1.17.0
typing-extensions==4.13.2
tzdata==2025.2
url-normalize==2.2.1
urllib3==2.2.3
urllib3-future==2.13.900
wassima==1.2.2
